pyproject.toml
[tool.poetry]
name = "app_for_analyze_bank_transactions"
version = "0.1.0"
description = ""
authors = ["FireBoy48 <gert48750@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pandas = "^2.3.0"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.0"
logger = "^1.4"
pytest = "^8.4.0"
pytest-cov = "^6.2.1"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.1"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
